FEATURE: rename folder src/data to src/mockData, and update other components based on rename
