+++ src/pages/MovieBookingPage/MovieBookingPage.tsx
@@ -1,5 +1,5 @@
 import { useParams } from "react-router-dom";
-import { MovieData } from "../../data/movieData";
+import { MovieData } from "../../mockData/movieData";
 import type { MovieDetails } from "../../types/interface";
 import { MovieBook } from "../../components/MovieBooking";
 import { PageNotFound } from "../../components/ui/PageNotFound";
